Handle runout in runout.cpp
